Move test folder logic into TestSystem
